login, register working. added wrapper for login and register added data service added model for user

+++ src/app/services/auth.service.ts
@@ -1,73 +1,60 @@
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs';
+import { User } from '../interface/user';
+import { DataService } from './data.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
 
-  users: any[] = [{
-    name: 'Yosi',
-    email: 'test@test.com',
-    password: '123456',
-    age: 26,
-    credentials: 'test@test.com-1234abcd'
-  },
-  {
-    email: 'test2@test.com',
-    password: '123456'
-  },
-  {
-    email: 'test1@test.com',
-    password: '123456'
-  }]
-
-
-  private _successfulAuth = new Subject<boolean>();
+  private _successfulAuth = new Subject<User>();
   authSuccess$ = this._successfulAuth.asObservable();
 
-  constructor() { }
+  constructor(private dataService: DataService) { }
   
   login(formValues: { email: string, password: string }) {
-    
     const { email, password } = formValues;
-
-    let success = false;
-
-    this.users.map((user) => {
-      if (user.email === email && user.password === password) {
-        this.setToken(email);
-        success = true
-      }
+    return new Promise ((resolve, reject) => {
+      let resolved = false
+      setTimeout(() => {
+        this.dataService.getUsers.map((user: User) => {
+          if (user.email === email && user.password === password) {
+            this.setToken(user);
+            resolve(true)
+            resolved = true
+          }
+        })
+        if (!resolved) reject()
+      }, 2500)
     })
-    if (!success) throw 'Email or password is wrong';
   }
 
-  private setToken(param: string) {
-    localStorage.setItem('credentials', `${param}-1234abcd`);
-    this._successfulAuth.next(true);
+  private setToken(user: User) {
+    this.dataService.setTokenOnUser = user.email;
+    this._successfulAuth.next(user);
   }
 
-  getToken() {
+  getLocalToken(): string | null  {
     return localStorage.getItem('credentials')
   }
 
-  checkMatchingCredentials() {
-    const creds = this.getToken();
-    let returnUser = null;
-    if (creds) {
-      this.users.map(user => {
-        if (user.credentials === creds) {
-          returnUser = user
-        }
-      })
+  getUserForLocalToken(): User | null {
+    let user = null;
+    const localToken = this.getLocalToken();
+    if (localToken) {
+      user = this.dataService.checkToken(localToken)
     }
-    return returnUser
+    return user
   }
 
   logout() {
-    localStorage.clear();
-    this._successfulAuth.next(false);
+    const localToken = this.getLocalToken();
+    if (localToken) {
+      localStorage.clear();
+      this.dataService.removeToken(localToken)
+    }
+    this._successfulAuth.next(undefined);
   }
 
 }
